refactor: moved login to separate router and controller
